// src/api/models/XcpPatches.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type XcpPatches = {
    version: string;
    url: string;
    size: number;
    release: string;
    name: string;
    license: string;
    description: string;
    changelog?: {
        description: string;
        date: number;
        author: string;
    };
};